new wrap for components
